+++ wemake_python_styleguide/violations/annotations.py
@@ -17,11 +17,13 @@ Summary
    :nosignatures:
 
    LiteralNoneViolation
+   NestedAnnotationsViolation
 
 Annotation checks
 ------------------
 
 .. autoclass:: LiteralNoneViolation
+.. autoclass:: NestedAnnotationsViolation
 
 """
 
@@ -59,3 +61,33 @@ class LiteralNoneViolation(ASTViolation):
 
     code = 701
     error_template = 'Found useless `Literal[None]` typing annotation'
+
+
+@final
+class NestedAnnotationsViolation(ASTViolation):
+    """
+    Forbids use of nested Literal and Union Annotation.
+
+    Reasoning:
+        There is no need to nest certain annotations of the same type.
+        They are exactly equivalent to the flattened version.
+        Use the flattened version for consistency.
+
+    Solution:
+        Flatten consecutively nested ``typing.Literal`` and ``typing.Union``.
+
+    Example::
+        # Correct:
+        Literal[1, 2, 3, "foo", 5, None]
+        Union[Union[int, str], float]
+
+        # Wrong:
+        Literal[Literal[Literal[1, 2, 3], "foo"], 5, None]
+        Union[int, str, float]
+
+    .. versionadded:: 0.13.0
+
+    """
+
+    error_template = 'Found redundant nested typing annotation'
+    code = 702