+++ CHANGELOG.md
@@ -9,13 +9,13 @@ We used to have incremental versioning before `0.1.0`.
 ### Features
 
 - Forbids using `Literal[None]` in function annotations
+- Forbids using nested `typing.Literal`, `typing.Union` and `typing.Annotated`
 - Forbids use of vague import names (e.g. `from json import loads`)
 - Make ``OveruseOfNoqaCommentViolation`` configurable (`--max-noqa-comments`)
 
 ### Bugfixes
 
 - Fixes ``ImplicitElifViolation`` false positives on a specific edge cases.
-
 
 ## 0.13.0